requirements.txt
python-binance==1.0.36
typer[all]==0.9.0
python-dotenv==1.0.0
